requirements.txt
viam-sdk
inference-sdk
Pillow
